Conflicts:
	app/views/data_files/index.html.haml

+++ app/controllers/cart_items_controller.rb
@@ -3,8 +3,8 @@ class CartItemsController < ApplicationController
   # GET /cart_items
   # GET /cart_items.json
   def index
-    @cart_items = CartItem.all
-
+    @cart_items = current_user.cart_items
+    session[:back]= request.referer
     respond_to do |format|
       format.html # index.html.erb
       format.json { render json: @cart_items }
@@ -30,32 +30,35 @@ class CartItemsController < ApplicationController
   # POST /cart_items
   # POST /cart_items.json
   def create
-      if params[:add_all] == 'true'
-        add_all
-      else
-        add_single
-      end
+    if params[:add_all] == 'true'
+      add_all
+    else
+      add_single
+    end
   end
 
   def add_single
+    session[:return_to]= request.referer
     @data_file = DataFile.find(params[:data_file_ids])
     @cart_item = current_user.cart_items.build
     @cart_item.data_file= @data_file
 
     respond_to do |format|
       if @cart_item.save
-        format.html { redirect_to data_file_path(@data_file),
-          notice: 'File was successfully added to cart.' }
+        format.html { redirect_to session[:return_to]||data_files_path,
+                                  notice: 'File was successfully added to cart.' }
         format.js { }
       else
-        format.html { redirect_to data_file_path(@data_file),
-          notice: 'File could not be added to cart.' }
+        format.html { redirect_to session[:return_to]||data_files_path,
+                                  notice: 'File could not be added to cart.' }
+        format.js { }
       end
     end
   end
 
   def add_all
     count = 0
+    session[:return_to]= request.referer
     params[:data_file_ids].each do |data_file_id|
       @data_file = DataFile.find(data_file_id)
       unless current_user.data_file_in_cart?(@data_file)
@@ -66,45 +69,38 @@ class CartItemsController < ApplicationController
       end
     end
     respond_to do |format|
-      format.html { redirect_to data_files_path(@data_file),
-          notice: "#{count} files were added to your cart." }
-      format.js { }
+      format.html {  redirect_to session[:return_to]||data_files_path,
+                                 notice: "#{count} files were added to your cart." }
     end
   end
 
   # DELETE /cart_items/1
   # DELETE /cart_items/1.json
   def destroy
+    session[:return_to]= request.referer
     @cart_item = CartItem.find(params[:id])
     @cart_item.destroy
 
     respond_to do |format|
-      format.html { redirect_to cart_items_url }
+      format.html { redirect_to session[:return_to], notice: "File was successfully removed from cart." }
       format.json { head :ok }
     end
   end
 
   def destroy_all
-    CartItem.where('user_id' == current_user.id).each do |cart_item|
-      unless cart_item.nil?
-        cart_item.destroy
+    session[:return_to]= request.referer
+    if current_user.data_files.empty?
+      redirect_to(data_files_path, :notice => "Your cart is empty.")
+    else
+      CartItem.where('user_id' == current_user.id).each do |cart_item|
+        unless cart_item.nil?
+          cart_item.destroy
+        end
       end
-    end
-    respond_to do |format|
-      format.html { redirect_to data_files_path, notice: 'Your cart was cleared.' }
-      format.js {  }
-    end
-  end
-
-   # DELETE /cart_items/1
-  # DELETE /cart_items/1.json
-  def download
-    @cart_item = CartItem.find(params[:id])
-    @cart_item.destroy
-
-    respond_to do |format|
-      format.html { redirect_to cart_items_url }
-      format.json { head :ok }
+      respond_to do |format|
+        format.html { redirect_to session[:return_to], notice: 'Your cart was cleared.' }
+        format.js {  }
+      end
     end
   end
 end
